react-spring package update // break fixes
Upgraded to react-spring v9, Keyframes was dropped from the package
which broke the Navbar component. Refactored Navbar component to be
functional component as well as use the hooks trail api in react-spring
v9. Need to use useTransition hook instead, will be cleaner. Known
Gatsby bug with react-spring also necessitated a modification to the
webpack config to set sideEffects to be true. Refs in index.js also
broke the build. Commented them out for now as I refactor the index
sequence for react-spring v9.